refactor(sign-up): use env executor to retrieve region at runtime
ref: MANAGER-9989
