+++ packages/manager/apps/sign-up/.eslintrc.json
@@ -1,7 +1,6 @@
 {
   "globals": {
     "__NODE_ENV__": true,
-    "__NG_APP_INJECTIONS__": true,
-    "__WEBPACK_REGION__": true
+    "__VERSION__": true
   }
 }
