+++ packages/manager/apps/sign-up/package.json
@@ -12,10 +12,7 @@
   "author": "OVH SAS",
   "main": "./src/index.js",
   "scripts": {
-    "build": "yarn build:eu && yarn build:ca && yarn build:us",
-    "build:ca": "webpack --env.production --env.region=\"CA\" && mv dist dist-CA",
-    "build:eu": "webpack --env.production && mv dist dist-EU",
-    "build:us": "webpack --env.production --env.region=\"US\" && mv dist dist-US",
+    "build": "webpack --env.production",
     "dev": "webpack-dev-server",
     "dev:watch": "yarn run dev",
     "start": "lerna exec --stream --scope='@ovh-ux/sign-up-app' --include-dependencies -- npm run build --if-present",
