+++ packages/manager/apps/sign-up/src/assets/img/logo/logo-ovhcloud-tm.png
