+++ packages/manager/apps/sign-up/src/assets/img/logo/logo-ovhcloudx2.png
