+++ packages/manager/apps/sign-up/src/assets/img/logo/logo-ovhcloudx3.png
