+++ packages/manager/apps/sign-up/webpack.config.js
@@ -4,10 +4,7 @@ const path = require('path');
 const webpackConfig = require('@ovh-ux/manager-webpack-config');
 const webpack = require('webpack');
 
-const REGION = process.env.REGION || 'EU';
-
 module.exports = (env = {}) => {
-  Object.assign(env, REGION ? { region: REGION } : {});
   const { config } = webpackConfig(
     {
       template: './src/index.html',
@@ -23,14 +20,7 @@ module.exports = (env = {}) => {
             to: 'flag-icon-css/flags/4x3',
           },
           {
-            from: path.resolve(
-              __dirname,
-              `src/assets/img/logo${
-                env.region && env.region.toUpperCase() === 'US'
-                  ? '/trademark'
-                  : '/classic'
-              }`,
-            ),
+            from: path.resolve(__dirname, 'src/assets/img/logo'),
             to: 'assets/img/logo',
           },
         ],
@@ -41,13 +31,11 @@ module.exports = (env = {}) => {
 
   // Extra config files
   const extras = glob.sync('./.extras/**/*.js');
-  const extrasRegion = glob.sync(`./.extras-${REGION}/**/*.js`);
 
   return merge(config, {
     entry: {
       main: './src/index.js',
       ...(extras.length > 0 ? { extras } : {}),
-      ...(extrasRegion.length > 0 ? { extrasRegion } : {}),
     },
     output: {
       path: path.join(__dirname, 'dist'),
@@ -55,12 +43,6 @@ module.exports = (env = {}) => {
     },
     plugins: [
       new webpack.DefinePlugin({
-        __NG_APP_INJECTIONS__: process.env.NG_APP_INJECTIONS
-          ? `'${process.env.NG_APP_INJECTIONS}'`
-          : 'null',
-        __WEBPACK_REGION__: process.env.REGION
-          ? `'${process.env.REGION.toUpperCase()}'`
-          : '"EU"',
         __NODE_ENV__: process.env.NODE_ENV
           ? `'${process.env.NODE_ENV}'`
           : '"development"',
